Move configuration work to library
