+++ internal/app/wwctl/configure/dhcp/root.go
@@ -7,17 +7,17 @@ import (
 var (
 	baseCmd = &cobra.Command{
 		DisableFlagsInUseLine: true,
-		Use:   "dhcp [OPTIONS]",
-		Short: "Manage and initialize DHCP",
+		Use:                   "dhcp [OPTIONS]",
+		Short:                 "Manage and initialize DHCP",
 		Long: "DHCP is a dependent service to Warewulf. This command will configure DHCP as defined\n" +
 			"in the warewulf.conf file.",
 		RunE: CobraRunE,
 	}
-	SetShow bool
+	setShow bool
 )
 
 func init() {
-	baseCmd.PersistentFlags().BoolVarP(&SetShow, "show", "s", false, "Show configuration (don't update)")
+	baseCmd.PersistentFlags().BoolVarP(&setShow, "show", "s", false, "Show configuration (don't update)")
 }
 
 // GetRootCommand returns the root cobra.Command for the application.
