+++ internal/app/wwctl/configure/nfs/root.go
@@ -5,17 +5,17 @@ import "github.com/spf13/cobra"
 var (
 	baseCmd = &cobra.Command{
 		DisableFlagsInUseLine: true,
-		Use:   "nfs [OPTIONS]",
-		Short: "Manage and initialize NFS",
+		Use:                   "nfs [OPTIONS]",
+		Short:                 "Manage and initialize NFS",
 		Long: "NFS is an optional dependent service of Warewulf, this tool will automatically\n" +
 			"configure NFS as per the configuration in the warewulf.conf file.",
 		RunE: CobraRunE,
 	}
-	SetShow bool
+	setShow bool
 )
 
 func init() {
-	baseCmd.PersistentFlags().BoolVarP(&SetShow, "show", "s", false, "Show configuration (don't update)")
+	baseCmd.PersistentFlags().BoolVarP(&setShow, "show", "s", false, "Show configuration (don't update)")
 }
 
 // GetRootCommand returns the root cobra.Command for the application.
