+++ internal/pkg/configure/dhcp_main_test.go
@@ -1,4 +1,4 @@
-package dhcp
+package configure
 
 import (
 	"path"
@@ -8,12 +8,12 @@ import (
 )
 
 func TestDhcpTemplateFile(t *testing.T) {
-	tests := []struct{
+	tests := []struct {
 		parameter string
-		expected string
+		expected  string
 	}{
 		{"", path.Join(buildconfig.SYSCONFDIR(), "warewulf/dhcp/default-dhcpd.conf")},
-		{"default", path.Join(buildconfig.SYSCONFDIR(),  "warewulf/dhcp/default-dhcpd.conf")},
+		{"default", path.Join(buildconfig.SYSCONFDIR(), "warewulf/dhcp/default-dhcpd.conf")},
 		{"static", path.Join(buildconfig.SYSCONFDIR(), "warewulf/dhcp/static-dhcpd.conf")},
 		{"/test/absolute/path.conf", "/test/absolute/path.conf"},
 	}
@@ -25,4 +25,3 @@ func TestDhcpTemplateFile(t *testing.T) {
 		}
 	}
 }
-
